requirements.txt
selenium==4.17.2
PyYAML==6.0.1
pandas==2.2.0